Fix bug where instantiated drop area would disappear
I had a lot of trouble with this bug. Apparently, when adding a child to the root of a tree in the _ready function, it produces an error because the tree is "busy setting up objects". To fix this, and adding children in the _ready function needs to be deferred